+++ src/components/badges/BadgeSection.tsx
@@ -1,41 +1,23 @@
-import { BodyText } from 'components/Text'
 import { ComponentChildren } from 'preact'
-import { useSnapshot } from 'valtio'
 import BadgeBlock from 'components/badges/BadgeBlock'
 import BadgesOwnedForContract from 'components/badges/BadgesOwnedForContract'
 import BaseProof from 'helpers/BaseProof'
-import Button from 'components/Button'
 import ChildrenProp from 'models/ChildrenProp'
 import Section from 'components/Section'
+import ShareToTwitterIfNeeded from 'components/badges/ShareToTwitterIfNeeded'
 import WalletStore from 'stores/WalletStore'
 import classnames, {
-  alignItems,
-  backgroundColor,
-  borderRadius,
   display,
-  flexDirection,
   gap,
   gridAutoRows,
-  gridColumn,
   gridTemplateColumns,
-  padding,
 } from 'classnames/tailwind'
-import shareToTwitter from 'helpers/shareToTwitter'
 
 const badgesList = classnames(
   display('grid'),
   gap('gap-2'),
   gridAutoRows('auto-rows-auto'),
   gridTemplateColumns('grid-cols-1', 'lg:grid-cols-2')
-)
-const wideBlock = classnames(
-  display('flex'),
-  flexDirection('flex-row'),
-  alignItems('items-center'),
-  borderRadius('rounded-lg'),
-  backgroundColor('bg-primary-dimmed'),
-  padding('lg:px-6', 'px-4', 'py-4'),
-  gridColumn('lg:col-span-2', 'col-span-1')
 )
 
 export default function ({
@@ -48,7 +30,6 @@ export default function ({
   proofs: BaseProof[]
 }) {
   if (minted.length === 0 && proofs.length === 0) return null
-  const { firstBadge } = useSnapshot(WalletStore)
 
   return (
     <Section title={title}>
@@ -59,6 +40,7 @@ export default function ({
             contractAddress={contractAddress}
           />
         ))}
+        <ShareToTwitterIfNeeded />
         {proofs.map((proof) => (
           <BadgeBlock
             onMinted={() => (WalletStore.firstBadge.minted = true)}
@@ -66,26 +48,6 @@ export default function ({
             proof={proof}
           />
         ))}
-        {firstBadge.minted && !firstBadge.notified && (
-          <div className={wideBlock}>
-            <BodyText bold fontPrimary>
-              You minted your first badge!
-            </BodyText>
-            <Button
-              type="secondary"
-              onClick={() => {
-                WalletStore.firstBadge.notified = true
-                shareToTwitter({
-                  text: 'Create zero knowledge proof and build your pseudonymous wallet with SealCred 🦭 sealcred.xyz',
-                })
-              }}
-              textMaxWidth
-              small
-            >
-              Share a Tweet
-            </Button>
-          </div>
-        )}
       </div>
     </Section>
   )
