+++ src/stores/WalletStore.ts
@@ -21,7 +21,7 @@ class WalletStore extends PersistableStore {
   walletsToNotifiedOfBeingDoxxed = {} as {
     [address: string]: boolean
   }
-  firstBadge = { minted: false, notified: false }
+  firstBadge = { minted: false, twitted: false }
 
   replacer = (key: string, value: unknown) => {
     const disallowList = ['account', 'walletLoading']
